values

+++ DES310_Game/Assets/Script/GameScripts/SaveGame.cs
@@ -6,7 +6,8 @@ using System.Runtime.Serialization.Formatters.Binary;
 
 public static class SaveGame
 {
-    public static void SaveGameData(int moneyData, int foodData, List<GameObject> grid)
+    //saves game data 
+    public static void SaveGameData(int moneyData, float foodData, List<GameObject> grid)
     {
         BinaryFormatter formatter = new BinaryFormatter();
 
@@ -21,6 +22,7 @@ public static class SaveGame
         stream.Close();
     }
 
+    //load game data
     public static SaveData LoadGameData()
     {
         string path = Application.persistentDataPath + "/saveData.SaveData";
